Work around Light Panel plugin causing Gazebo freezes
The light panel plugin creates a new object for the light on every update, which was causing Gazebo freezes. This works around that by only performing the update when the light color needs to change. It means that the pose of the lights doesn't update until the next color change.